+++ engine/compmgr/src/assist/reaction.js
@@ -510,18 +510,24 @@ module.exports = (...args) => {
               }
               let script = document.createElement("script");
               script.type = "text/javascript";
-              script.innerHTML = `var mjs=${JSON.stringify(
+              script.innerHTML = `var glib = {},library = {},sys = {},mjs = ${JSON.stringify(
                 import_mjs(mjs, params)
               )};`;
               if (Object.keys(injectionjs.variables).length > 0)
                 script.innerHTML += `var injectionjs=${JSON.stringify(
                   injectionjs.variables
                 )}`;
-              let libutils = document.createElement("script");
-              libutils.type = "module"; // Crucial for ES module
-              libutils.src = "/library/utils.js"; // For external module
 
-              document.getElementsByTagName("head")[0].appendChild(libutils);
+              [
+                "/library/utils.js",
+                "/library/engine/workflow.js",
+              ].map((value) => {
+                let libraries = document.createElement("script");
+                libraries.type = "module"; // Crucial for ES module
+                libraries.src = value; // For external module
+                document.getElementsByTagName("head")[0].appendChild(libraries);
+              });
+
               document.getElementsByTagName("head")[0].appendChild(script);
               let rtnimport_css = import_css(document, css, params);
               if (rtnimport_css) throw rtnimport_css;