+++ engine/deskelectronjs/src/server.js
@@ -68,7 +68,7 @@ module.exports = (...args) => {
         const { general, genernalexcludefile, logpath, share } = setting;
         const { savestore, store, verbose, ...setsession } =
           setting[setting.args.engine].session;
-        const { assets, atomic, reaction, utilities } = compmgr.assist;
+        const { assets, atomic, reaction, libraries } = compmgr.assist;
 
         try {
           // Setup server log
@@ -121,7 +121,7 @@ module.exports = (...args) => {
             serveStatic,
           });
 
-          await utilities(["utils"], {
+          await libraries(["utils", "engine/workflow"], {
             register: app.use,
             serveStatic,
             library,