+++ engine/workflow/src/index.js
@@ -1,0 +1,264 @@
+/**
+ * Copyright (c) 2025   Loh Wah Kiang
+ *
+ * openGauss is licensed under Mulan PSL v2.
+ * You can use this software according to the terms and conditions of the Mulan PSL v2.
+ * You may obtain a copy of Mulan PSL v2 at:
+ *
+ *          http://license.coscl.org.cn/MulanPSL2
+ *
+ * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
+ * EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
+ * MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
+ * See the Mulan PSL v2 for more details.
+ * -------------------------------------------------------------------------
+ */
+"use strict";
+
+/**
+ * The core of workflow engine module
+ * @module src_wfengine
+ */
+
+module.exports = (...args) => {
+  const [params, obj] = args;
+  const [pathname, curdir] = params;
+  const [library, sys, cosetting] = obj;
+
+  try {
+    const load = async (...args) => {
+      const [param] = args;
+      let htmlengine = {};
+      for (let [name, value] of Object.entries(param.load)) {
+        if (name.substring(0, 4) == "html") {
+          let folder = name.substring(4);
+          htmlengine[name] = {};
+          for (let [parent, child] of Object.entries(value)) {
+            for (let item of child) {
+              let fnpath = `${param.path}${parent}/${folder}/${item}`;
+              let fn = fnpath.split("/").pop().replace(".js", "");
+              fn = fn.replace(".", "-");
+              let { default: df } = await import(fnpath);
+              if (df) {
+                let { load, register, ...other } = df;
+                if (htmlengine[name][fn])
+                  htmlengine[name][fn] = {
+                    ...htmlengine[name][fn],
+                    ...other,
+                  };
+                else htmlengine[name][fn] = other;
+              }
+            }
+          }
+        }
+      }
+      return htmlengine;
+    };
+
+    const register = (...args) => {
+      const [params, obj] = args;
+      const [htmlengine, trigger] = params;
+      const { htmlcollection, htmlevent, htmllogicflow, htmlrender } =
+        htmlengine;
+      const { convtrigger, regevents } = obj;
+      const { mergeDeep } = library.utils;
+      let objfuncs = {};
+      for (let funcs of [htmllogicflow, htmlrender, htmlcollection])
+        objfuncs = mergeDeep(objfuncs, { ...funcs });
+
+      regevents(convtrigger(trigger), htmlevent);
+      return objfuncs;
+    };
+
+    const convtrigger = (...args) => {
+      const [trigger] = args;
+      const { getNestedObject, handler, jptr } = library.utils;
+      const html_objevents = handler.winevents;
+      let objevents = handler.winevents;
+      for (let [name, value] of Object.entries(trigger)) {
+        let fn = getNestedObject(html_objevents, name);
+        if (fn) {
+          let location = name.replaceAll(".", "/");
+          jptr.set(objevents, location, value);
+        }
+      }
+      return objevents;
+    };
+
+    const run = async (...args) => {
+      const [params, obj] = args;
+      const [event, task, showdata = true] = params;
+      const { htmlengine, objfuncs } = obj;
+      const {
+        datatype,
+        errhandler,
+        getNestedObject,
+        handler,
+        pick_arrayofobj,
+        pick_arrayobj2list,
+        serialize,
+      } = library.utils;
+      let output = handler.dataformat;
+
+      try {
+        let { htmlworkflow } = htmlengine;
+        let objkeys = Object.keys(htmlworkflow);
+        let fn, func, required;
+        let task_type = datatype(task);
+        if (task_type == "string") func = task;
+        else if (task_type == "object") {
+          let { taskname, required: req } = task;
+          func = taskname;
+          if (req) required = req;
+        }
+
+        const parallel = async (...args) => {
+          const [input, share] = args;
+          let process = [];
+          for (let objserialize of input) {
+            objserialize.func = objfuncs;
+            objserialize.share = share;
+            let rtn = await new serialize(
+              objserialize,
+              [library, sys],
+              showdata
+            );
+            process.push(rtn);
+            if (rtn.code != 0) break;
+          }
+          return process;
+        };
+        for (let item of objkeys) {
+          let getfn = getNestedObject(htmlworkflow, `${item}.${func}`);
+          if (getfn) {
+            fn = getfn;
+            break;
+          }
+        }
+        if (fn) {
+          let rtnprocess = [];
+          let inputs = fn(event, func);
+          if (inputs) {
+            let share = { shared: { htmlengine, convtrigger } };
+            if (required) share.shared = { ...share.shared, required };
+            if (datatype(inputs) == "object") inputs = [inputs];
+            let allexec = [];
+            for (let input of inputs) {
+              if (datatype(input) == "object") {
+                input.func = objfuncs;
+                input.share = share;
+                allexec.push(new serialize(input, [library, sys], showdata));
+              } else if (datatype(input) == "array")
+                allexec.push(parallel(input, share));
+            }
+            rtnprocess = await Promise.all(allexec);
+
+            let code = "";
+            for (let chkrtnproc of rtnprocess) {
+              if (datatype(chkrtnproc) == "array") {
+                let pickarrobj = pick_arrayofobj(chkrtnproc, ["code"]);
+                let chkcode = pick_arrayobj2list(pickarrobj, [
+                  "code",
+                ]).code.join("");
+                let validcode = "0".padStart(chkrtnproc.length, "0");
+                if (chkcode != validcode) code += `[${chkcode}],`;
+              } else {
+                if (chkrtnproc["code"] != 0)
+                  code += `[${chkrtnproc["code"].toString()}],`;
+              }
+            }
+
+            if (code != "") {
+              output.code = -1;
+              output.msg = code;
+            }
+            output.data = rtnprocess;
+          } else {
+            output.code = -1;
+
+            output.msg = `Error: Incomplete workflow in ${func} process`;
+          }
+        }
+      } catch (error) {
+        output = errhandler(error);
+      } finally {
+        return output;
+      }
+    };
+
+    const regevents = (...args) => {
+      const [param, objfuncs] = args;
+      const { utils } = library;
+      const { datatype, errhandler, getNestedObject, handler } = utils;
+      let output = handler.dataformat;
+      try {
+        for (let [, valobjevent] of Object.entries(param)) {
+          for (let [key, value] of Object.entries(valobjevent)) {
+            for (let [evt, fn] of Object.entries(value)) {
+              let qs = document.querySelectorAll(evt);
+              if (qs) {
+                for (let nodevalue of qs) {
+                  if (typeof fn === "function")
+                    nodevalue.addEventListener(key, fn);
+                  else if (datatype(fn) === "string") {
+                    let func = getNestedObject(objfuncs, fn);
+                    if (func) nodevalue.addEventListener(key, func);
+                  } else if (datatype(fn) === "object") {
+                    if (fn.attr) {
+                      for (let [attrkey, attrval] of Object.entries(fn.attr))
+                        nodevalue.setAttribute(attrkey, attrval);
+                    }
+                    let func = getNestedObject(objfuncs, fn.evt);
+                    if (func) nodevalue.addEventListener(key, func);
+                  }
+                }
+              }
+            }
+          }
+        }
+      } catch (error) {
+        output = errhandler(error);
+      } finally {
+        return output;
+      }
+    };
+
+    const loadlib = async (...args) => {
+      try {
+        let [param] = args;
+        let libraries = {};
+        for (let item of param) {
+          let fn = item.split("/").pop().replace(".js", "");
+          let [, atomic] = item.split("/");
+          if (atomic == "atomic") {
+            if (!libraries[atomic]) libraries[atomic] = {};
+          }
+          fn = fn.replace(".", "-");
+          let { default: df, ...otherlib } = await import(item);
+          if (Object.keys(otherlib).length > 0) {
+            if (atomic !== "atomic") glib[fn] = { ...df, ...otherlib };
+            else libraries[atomic][fn] = { ...df, ...otherlib };
+          } else {
+            if (atomic !== "atomic") glib[fn] = df;
+            else libraries[atomic][fn] = df;
+          }
+        }
+        return libraries;
+      } catch (error) {
+        console.log(error);
+      }
+    };
+
+    return {
+      convtrigger,
+      load,
+      loadlib,
+      regevents,
+      register,
+      run,
+      excluded: ["excluded"],
+    };
+  } catch (error) {
+    return library.utils.errhandler(error);
+  }
+};